Various fixes for use_target_ruby.
Add BP_DEGUG flag and centralized shell command execution using open3
to get good quality debugging output on all of the shelling out fun
required for use_target_ruby.

`gem_home` derives from target Ruby `Gem.default_dir`, not
`Bundler::Settings#path` since the latter will defer to `GEM_HOME` and
that's not what we want. We also don't need to care about any local
Bundler config for `path` - bundler-patch can be installed in the
"global" Ruby gems and still work.

Fix the gem install command, to install Bundler patch into target Ruby.

+++ spec/bundler/integration/use_target_ruby_spec.rb
@@ -1,5 +1,3 @@
-# TODO: maybe don't need a separate file here? meh.
-
 require_relative '../../spec_helper'
 
 require 'tmpdir'
@@ -10,7 +8,7 @@ describe 'integration tests' do
   end
 
   after do
-    FileUtils.remove_entry_secure(@tmp_dir)
+    FileUtils.remove_entry_secure(@tmp_dir) if File.exist?(@tmp_dir)
   end
 
   def gemfile_create(ruby_version)
@@ -41,8 +39,7 @@ describe 'integration tests' do
       # in the fixture code to do the lock command against a declared older Ruby.
       glf = GemfileLockFixture.new(dir: @tmp_dir,
                                    gems: {rack: '~> 1.4.1', addressable: '2.1.1'},
-                                   # TODO - need programmatic way to work with an old ruby and make sure TRAVIS has it installed
-                                   ruby_version: '2.3.3')
+                                   ruby_version: '2.1.10')
       glf.create_gemfile
       bf = glf.bundler_fixture
 
@@ -60,6 +57,28 @@ describe 'integration tests' do
       output.should match /addressable 2\.1\.1/
     end
 
-    it 'with different ruby bundle config install path'
+    it 'with different ruby bundle config install path' do
+      # Only the Gemfile is created here, with no lock file, because it won't work
+      # in the fixture code to do the lock command against a declared older Ruby.
+      glf = GemfileLockFixture.new(dir: @tmp_dir,
+                                   gems: {rack: '~> 1.4.1', addressable: '2.1.1'},
+                                   ruby_version: '2.1.10')
+      glf.create_gemfile
+      bf = glf.bundler_fixture
+      bf.create_config(path: 'local-path')
+
+      output = nil
+      with_clean_env do
+        output = bundler_patch(gems_to_update: ['rack'],
+                               gemfile: File.join(@tmp_dir, 'Gemfile'),
+                               use_target_ruby: true)
+      end
+
+      bf.lockfile_spec_version('rack').should == '1.4.7'
+      bf.lockfile_spec_version('addressable').should == '2.1.1'
+
+      output.should match /rack 1\.4\.7/
+      output.should match /addressable 2\.1\.1/
+    end
   end
 end
